resourcemanager/tags: refactoring the expand/flatten/validate into their own package
This means that commonschema is just for schema
