+++ resourcemanager/tags/flatten.go
@@ -1,0 +1,18 @@
+package tags
+
+import (
+	"github.com/hashicorp/go-azure-helpers/lang/pointer"
+)
+
+func Flatten(input *map[string]string) map[string]*string {
+	output := make(map[string]*string)
+	if input == nil {
+		return output
+	}
+
+	for k, v := range *input {
+		output[k] = pointer.ToStringPointer(v)
+	}
+
+	return output
+}
